requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
mss>=6.1.0
Pillow>=8.0.0
pyautogui>=0.9.50